Refactor to allow for multiple parsers
